backend/test_signed_sync.py
#!/usr/bin/env python
"""
Test script for Step2Win anti-cheat signed sync requests.
Simulates mobile app sending HMAC-signed step data to backend.
"""

import os
import sys
import django
import json
import hmac
import hashlib
import uuid
from datetime import datetime

# Setup Django
os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'step2win.settings')
sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
django.setup()

from django.contrib.auth import get_user_model  # noqa: E402
from rest_framework.test import APIClient  # noqa: E402
from rest_framework_simplejwt.tokens import RefreshToken  # noqa: E402
from django.conf import settings  # noqa: E402

User = get_user_model()


def get_access_token(user) -> str:
    return str(RefreshToken.for_user(user).access_token)


def generate_hmac_signature(user_id: int, body: str, secret: str) -> tuple[str, str, str]:
    """
    Generate HMAC signature as frontend would.
    
    Returns: (signature, timestamp, idempotency_key)
    """
    timestamp = str(int(datetime.now().timestamp()))
    body_sha = hashlib.sha256(body.encode()).hexdigest()
    
    msg = f"{user_id}:{timestamp}:{body_sha}"
    sig = hmac.new(secret.encode(), msg.encode(), hashlib.sha256).hexdigest()
    
    idempotency_key = str(uuid.uuid4())
    
    return sig, timestamp, idempotency_key


def test_valid_signed_sync():
    """Test: Valid signed request should be accepted."""
    print("\n" + "="*60)
    print("TEST 1: Valid Signed Sync Request")
    print("="*60)
    
    # Setup test user
    user, created = User.objects.get_or_create(
        username='test_signed_sync_user',
        defaults={'email': 'test@signed.test', 'first_name': 'Test'}
    )
    print(f"✓ User: {user.username} (ID: {user.id})")
    
    # Prepare sync data
    sync_body = json.dumps({
        "steps": 8500,
        "date": "2026-03-07",
        "distance_km": 5.2,
        "calories": 425,
        "active_minutes": 45
    })
    
    # Generate signature
    app_secret = settings.APP_SIGNING_SECRET
    sig, timestamp, idempotency_key = generate_hmac_signature(user.id, sync_body, app_secret)
    
    print(f"✓ Generated HMAC signature: {sig[:16]}...")
    print(f"✓ Timestamp: {timestamp}")
    print(f"✓ Idempotency Key: {idempotency_key}")
    
    access_token = get_access_token(user)
    client = APIClient()
    
    response = client.post(
        '/api/steps/sync/',
        data=sync_body,
        content_type='application/json',
        HTTP_AUTHORIZATION=f'Bearer {access_token}',
        HTTP_X_APP_SIGNATURE=sig,
        HTTP_X_TIMESTAMP=timestamp,
        HTTP_X_IDEMPOTENCY_KEY=idempotency_key,
    )
    
    print(f"\n✓ Response Status: {response.status_code}")
    
    if response.status_code in [200, 201]:
        data = response.json()
        print(f"✓ Approved Steps: {data.get('approved_steps', 'N/A')}")
        print(f"✓ Trust Score: {data.get('trust_score', 'N/A')}")
        print(f"✓ Trust Status: {data.get('trust_status', 'N/A')}")
        print(f"✓ Flags Raised: {data.get('flags_raised', 0)}")
        print("\n✅ TEST PASSED: Valid sync accepted with trust score visible")
        return True
    else:
        print(f"❌ Response: {response.content}")
        print("❌ TEST FAILED: Expected 200/201, got", response.status_code)
        return False


def test_invalid_signature():
    """Test: Invalid signature should be rejected."""
    print("\n" + "="*60)
    print("TEST 2: Invalid Signature (Should Be Rejected)")
    print("="*60)
    
    user, _ = User.objects.get_or_create(
        username='test_invalid_sig_user',
        defaults={'email': 'test@invalid.test', 'first_name': 'Test'}
    )
    print(f"✓ User: {user.username} (ID: {user.id})")
    
    sync_body = json.dumps({"steps": 5000, "date": "2026-03-07"})
    
    # Generate WRONG signature (use wrong secret)
    wrong_secret = "totally-wrong-secret"
    sig, timestamp, idempotency_key = generate_hmac_signature(user.id, sync_body, wrong_secret)
    
    print("✓ Generated BAD signature (wrong secret)")
    
    access_token = get_access_token(user)
    client = APIClient()
    
    response = client.post(
        '/api/steps/sync/',
        data=sync_body,
        content_type='application/json',
        HTTP_AUTHORIZATION=f'Bearer {access_token}',
        HTTP_X_APP_SIGNATURE=sig,
        HTTP_X_TIMESTAMP=timestamp,
        HTTP_X_IDEMPOTENCY_KEY=idempotency_key,
    )
    
    print(f"\n✓ Response Status: {response.status_code}")
    
    if response.status_code == 401 or response.status_code == 403:
        print(f"✓ Response: {response.content[:100]}")
        print("✅ TEST PASSED: Invalid signature rejected as expected")
        return True
    else:
        print(f"❌ Expected 401/403, got {response.status_code}")
        print("❌ TEST FAILED: Invalid signature was not rejected")
        return False


def test_idempotency():
    """Test: Duplicate request (same idempotency key) should return cached response."""
    print("\n" + "="*60)
    print("TEST 3: Idempotency Protection (Duplicate Request)")
    print("="*60)
    
    user, _ = User.objects.get_or_create(
        username='test_idempotency_user',
        defaults={'email': 'test@idempotency.test', 'first_name': 'Test'}
    )
    print(f"✓ User: {user.username} (ID: {user.id})")
    
    sync_body = json.dumps({
        "steps": 7000,
        "date": "2026-03-07",
        "distance_km": 4.3,
        "calories": 350,
        "active_minutes": 38
    })
    
    app_secret = settings.APP_SIGNING_SECRET
    sig, timestamp, idempotency_key = generate_hmac_signature(user.id, sync_body, app_secret)
    
    print(f"✓ Generated signature and idempotency key: {idempotency_key[:8]}...")
    
    access_token = get_access_token(user)
    client = APIClient()
    
    # First request
    response1 = client.post(
        '/api/steps/sync/',
        data=sync_body,
        content_type='application/json',
        HTTP_AUTHORIZATION=f'Bearer {access_token}',
        HTTP_X_APP_SIGNATURE=sig,
        HTTP_X_TIMESTAMP=timestamp,
        HTTP_X_IDEMPOTENCY_KEY=idempotency_key,
    )
    
    print(f"✓ First Request Status: {response1.status_code}")
    
    # Second request with SAME idempotency key
    response2 = client.post(
        '/api/steps/sync/',
        data=sync_body,
        content_type='application/json',
        HTTP_AUTHORIZATION=f'Bearer {access_token}',
        HTTP_X_APP_SIGNATURE=sig,
        HTTP_X_TIMESTAMP=timestamp,
        HTTP_X_IDEMPOTENCY_KEY=idempotency_key,  # ← Same key
    )
    
    print(f"✓ Duplicate Request Status: {response2.status_code}")
    
    if response2.status_code == 409:
        print("✓ Got 409 Conflict (duplicate detected)")
        print("✅ TEST PASSED: Idempotency protection working")
        return True
    elif response2.status_code == response1.status_code:
        print("✓ Got same status code (may indicate cache)")
        print("⚠️  TEST PARTIAL: Got same response (cache or no check)")
        return True
    else:
        print(f"❌ Expected 409 or {response1.status_code}, got {response2.status_code}")
        print("❌ TEST FAILED: Idempotency not working as expected")
        return False


def test_stale_timestamp():
    """Test: Stale timestamp (>5 min old) should be rejected."""
    print("\n" + "="*60)
    print("TEST 4: Stale Timestamp (Should Be Rejected)")
    print("="*60)
    
    user, _ = User.objects.get_or_create(
        username='test_stale_ts_user',
        defaults={'email': 'test@stale.test', 'first_name': 'Test'}
    )
    print(f"✓ User: {user.username} (ID: {user.id})")
    
    sync_body = json.dumps({"steps": 6000, "date": "2026-03-07"})
    
    app_secret = settings.APP_SIGNING_SECRET
    
    # Generate signature with OLD timestamp (10 minutes ago)
    old_timestamp = str(int(datetime.now().timestamp()) - 600)  # 10 min old
    body_sha = hashlib.sha256(sync_body.encode()).hexdigest()
    msg = f"{user.id}:{old_timestamp}:{body_sha}"
    sig = hmac.new(app_secret.encode(), msg.encode(), hashlib.sha256).hexdigest()
    
    print("✓ Generated signature with 10-minute-old timestamp")
    
    access_token = get_access_token(user)
    client = APIClient()
    
    response = client.post(
        '/api/steps/sync/',
        data=sync_body,
        content_type='application/json',
        HTTP_AUTHORIZATION=f'Bearer {access_token}',
        HTTP_X_APP_SIGNATURE=sig,
        HTTP_X_TIMESTAMP=old_timestamp,
        HTTP_X_IDEMPOTENCY_KEY=str(uuid.uuid4()),
    )
    
    print(f"\n✓ Response Status: {response.status_code}")
    
    if response.status_code in [401, 403]:
        print(f"✓ Response: {response.content[:100]}")
        print("✅ TEST PASSED: Stale timestamp rejected")
        return True
    else:
        print(f"❌ Expected 401/403, got {response.status_code}")
        print("⚠️  TEST FAILED: Stale timestamp was not rejected")
        return False


def main():
    """Run all tests."""
    print("\n" + "#"*60)
    print("# Step2Win Anti-Cheat Signed Sync Test Suite")
    print("#"*60)
    print(f"\nAPP_SIGNING_SECRET: {settings.APP_SIGNING_SECRET[:16]}...")
    print(f"Django DEBUG: {settings.DEBUG}")
    
    tests = [
        test_valid_signed_sync,
        test_invalid_signature,
        test_idempotency,
        test_stale_timestamp,
    ]
    
    results = []
    for test_func in tests:
        try:
            result = test_func()
            results.append(result)
        except Exception as e:
            print(f"\n❌ TEST FAILED WITH EXCEPTION: {e}")
            import traceback
            traceback.print_exc()
            results.append(False)
    
    # Summary
    print("\n" + "#"*60)
    print("# Test Results Summary")
    print("#"*60)
    passed = sum(results)
    total = len(results)
    print(f"\n✓ {passed}/{total} tests passed")
    
    if passed == total:
        print("\n🎉 All tests passed! Anti-cheat signing is working correctly.")
        return 0
    else:
        print("\n⚠️  Some tests failed. Review output above for details.")
        return 1


if __name__ == '__main__':
    sys.exit(main())
