backend/requirements.txt
Django==5.0.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-defender==0.9.8
drf-spectacular==0.27.0
psycopg2-binary==2.9.10
celery==5.3.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
sqlalchemy==2.0.40
channels==4.1.0
sentry-sdk==1.40.0
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==11.1.0
setuptools==80.9.0
requests==2.31.0
mammoth==1.6.0
PyPDF2==3.0.1
